refactored w/ new file
